requirements.txt
aenum==3.1.16
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
attrs==25.3.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
comm==0.2.2
debugpy==1.8.5
decorator==5.1.1
Deprecated==1.2.18
Django==5.1.2
et_xmlfile==2.0.0
executing==2.1.0
fastapi==0.115.6
Flask==3.1.0
frozenlist==1.7.0
future==1.0.0
gunicorn==23.0.0
idna==3.10
ipykernel==6.29.5
ipython==8.27.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.5
jupyter_client==8.6.3
jupyter_core==5.7.2
line-bot-sdk==3.17.1
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
multidict==6.4.4
nest-asyncio==1.6.0
numpy==1.26.4
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
parso==0.8.4
peppercorn==0.6
pexpect==4.9.0
platformdirs==4.3.6
prompt_toolkit==3.0.47
propcache==0.3.2
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyzmq==26.2.0
requests==2.32.4
sampleproject==3.0.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
stack-data==0.6.3
starlette==0.41.3
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.4.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
xlrd==2.0.1
yarl==1.20.1
